requirements.txt
BeautifulSoup==3.2.1
cffi==1.6.0
characteristic==14.3.0
cryptography==1.3.2
cssselect==0.9.1
enum34==1.1.6
goose-extractor==1.0.25
idna==2.1
ipaddress==1.0.16
jieba==0.38
lxml==3.4.4
nltk==3.2.1
Pillow==3.2.0
pyasn1==0.1.9
pyasn1-modules==0.0.5
pycparser==2.14
pyOpenSSL==16.0.0
queuelib==1.2.2
Scrapy==0.24.6
service-identity==14.0.0
six==1.10.0
textblob==0.11.1
Twisted==15.2.1
w3lib==1.11.0
zope.interface==4.1.2
